requirements.txt
pydub
ffmpeg
matplotlib
requests-futures
